started splitting out a connection manager class
