+++ dbt/adapters/base/__init__.py
@@ -1,2 +1,3 @@
 from dbt.adapters.base.relation import BaseRelation
+from dbt.adapters.base.connections import BaseConnectionManager
 from dbt.adapters.base.impl import BaseAdapter
