+++ dbt/adapters/default/__init__.py
@@ -1,0 +1,8 @@
+from dbt.adapters.default.impl import DefaultAdapter, ConnectionManager, \
+    CommonSQLAdapter
+from dbt.adapters.default.relation import DefaultRelation
+
+__all__ = [
+    'DefaultAdapter',
+    'DefaultRelation',
+]
