+++ dbt/adapters/sql/impl.py
@@ -26,14 +26,30 @@ class SQLAdapter(BaseAdapter):
         - date_function
         - get_existing_schemas
         - list_relations_without_caching
-        - cancel_connection
-        - get_status
         - get_columns_in_relation_sql
 
     """
     config_functions = BaseAdapter.config_functions[:] + [
         'add_query',
     ]
+
+    def add_query(self, sql, model_name=None, auto_begin=True, bindings=None,
+                  abridge_sql_log=False):
+        """Add a query to the current transaction. A thin wrapper around
+        ConnectionManager.add_query.
+
+        :param str sql: The SQL query to add
+        :param Optional[str] model_name: The name of the connection the
+            transaction is on
+        :param bool auto_begin: If set and there is no transaction in progress,
+            begin a new one.
+        :param Optional[List[object]]: An optional list of bindings for the
+            query.
+        :param bool abridge_sql_log: If set, limit the raw sql logged to 512
+            characters
+        """
+        return self.connections.add_query(sql, model_name, auto_begin,
+                                          bindings, abridge_sql_log)
 
     @classmethod
     def convert_text_type(cls, agate_table, col_idx):
@@ -76,75 +92,6 @@ class SQLAdapter(BaseAdapter):
     @classmethod
     def is_cancelable(cls):
         return True
-
-    @abc.abstractmethod
-    def cancel_connection(connection):
-        """Cancel the given connection.
-
-        :param Connection connection: The connection to cancel.
-        """
-        raise dbt.exceptions.NotImplementedException(
-            '`cancel_connection` is not implemented for this adapter!'
-        )
-
-    def cancel_open_connections(self):
-        global connections_in_use
-
-        for name, connection in connections_in_use.items():
-            if name == 'master':
-                continue
-
-            self.cancel_connection(connection)
-            yield name
-
-    def add_query(self, sql, model_name=None, auto_begin=True,
-                  bindings=None, abridge_sql_log=False):
-        connection = self.get_connection(model_name)
-        connection_name = connection.name
-
-        if auto_begin and connection.transaction_open is False:
-            self.begin(connection_name)
-
-        logger.debug('Using {} connection "{}".'
-                     .format(self.type(), connection_name))
-
-        with self.exception_handler(sql, connection_name):
-            if abridge_sql_log:
-                logger.debug('On %s: %s....', connection_name, sql[0:512])
-            else:
-                logger.debug('On %s: %s', connection_name, sql)
-            pre = time.time()
-
-            cursor = connection.handle.cursor()
-            cursor.execute(sql, bindings)
-
-            logger.debug("SQL status: %s in %0.2f seconds",
-                         self.get_status(cursor), (time.time() - pre))
-
-            return connection, cursor
-
-    @abstractclassmethod
-    def get_status(cls, cursor):
-        """Get the status of the cursor.
-
-        :param cursor: A database handle to get status from
-        :return: The current status
-        :rtype: str
-        """
-        raise dbt.exceptions.NotImplementedException(
-            '`get_status` is not implemented for this adapter!'
-        )
-
-    def execute(self, sql, model_name=None, auto_begin=False,
-                fetch=False):
-        self.get_connection(model_name)
-        _, cursor = self.add_query(sql, model_name, auto_begin)
-        status = self.get_status(cursor)
-        if fetch:
-            table = self.get_result_from_cursor(cursor)
-        else:
-            table = dbt.clients.agate_helper.empty_table()
-        return status, table
 
     def expand_column_types(self, goal, current, model_name=None):
         reference_columns = {
